Invalidate `sys.modules` cache after each test
